+++ quic/codec/PacketNumberCipher.h
@@ -10,7 +10,6 @@
 
 #include <folly/Optional.h>
 #include <folly/io/Cursor.h>
-#include <folly/ssl/OpenSSLPtrTypes.h>
 
 namespace quic {
 
@@ -88,31 +87,4 @@ class PacketNumberCipher {
       uint8_t packetNumLengthMask) const;
 };
 
-class Aes128PacketNumberCipher : public PacketNumberCipher {
- public:
-  ~Aes128PacketNumberCipher() override = default;
-
-  void setKey(folly::ByteRange key) override;
-
-  HeaderProtectionMask mask(folly::ByteRange sample) const override;
-
-  size_t keyLength() const override;
-
- private:
-  folly::ssl::EvpCipherCtxUniquePtr encryptCtx_;
-};
-
-class Aes256PacketNumberCipher : public PacketNumberCipher {
- public:
-  ~Aes256PacketNumberCipher() override = default;
-
-  void setKey(folly::ByteRange key) override;
-
-  HeaderProtectionMask mask(folly::ByteRange sample) const override;
-
-  size_t keyLength() const override;
-
- private:
-  folly::ssl::EvpCipherCtxUniquePtr encryptCtx_;
-};
 } // namespace quic
